+++ src/Cmd/Override.hs
@@ -6,20 +6,25 @@ import Common.System
 
 import Bodhi
 import Branches
+import Cmd.WaitRepo (waitrepoCmd)
 import Git
 import Koji
 import Package
 
 -- FIXME option to expire (all) overrides
-overrideCmd :: Bool -> Maybe Int -> [Branch] -> [String] -> IO ()
-overrideCmd dryrun mduration brs pkgs =
-  withPackageByBranches (Just False) cleanGitFetchActive AnyNumber overrideBranch (Branches brs, pkgs)
+overrideCmd :: Bool -> Maybe Int -> Bool -> (BranchesReq, [String]) -> IO ()
+overrideCmd dryrun mduration nowait breqpkgs = do
+  unless nowait $
+    putStrLn "Overriding"
+  withPackageByBranches (Just False) cleanGitFetchActive AnyNumber overrideBranch breqpkgs
+  unless nowait $ do
+    putStrLn "Waiting"
+    waitrepoCmd dryrun breqpkgs
   where
     overrideBranch :: Package -> AnyBranch -> IO ()
     overrideBranch _ (OtherBranch _) =
       error' "override only defined for release branches"
     overrideBranch pkg rbr@(RelBranch br) = do
-      putPkgBrnchHdr pkg br
       gitSwitchBranch rbr
       let spec = packageSpec pkg
       nvr <- pkgNameVerRel' br spec
@@ -28,4 +33,3 @@ overrideCmd dryrun mduration brs pkgs =
       unless (any (`elem` tags) [show br, show br ++ "-updates", show br ++ "-override"]) $
         unlessM (checkAutoBodhiUpdate br) $ do
         bodhiCreateOverride dryrun mduration nvr
-        kojiWaitRepo dryrun (branchTarget br) nvr